pyproject.toml
[tool.poetry]
name = "stimkysticker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
telethon = "^1.29.3"
pillow = "^10.0.0"
brother-ql = "^0.9.4"
attrs = "^23.1.0"
pyyaml = "^6.0.1"
loguru = "^0.7.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
